separated normalize functions to a method file
