+++ App.js
@@ -1,12 +1,12 @@
 import {Login, SignUp, MyComplaints} from './src/containers';
 
 import React from 'react';
-import {View} from 'react-native';
 import {useGoogleConfiguration} from './src/library/hooks';
 import {withRedux} from './src/library/redux';
 
 import {NavigationContainer} from '@react-navigation/native';
 import {createStackNavigator} from '@react-navigation/stack';
+
 const Stack = createStackNavigator();
 
 const App = () => {
