+++ src/containers/MyComplaints/styles/index.js
@@ -1,19 +1,6 @@
 import {StyleSheet} from 'react-native';
 import {colors} from '../../../library/styles/vars';
-import {Dimensions, Platform, PixelRatio} from 'react-native';
-
-const {width: SCREEN_WIDTH, height: SCREEN_HEIGHT} = Dimensions.get('window');
-
-const scale = SCREEN_WIDTH / 370;
-
-const normalize = size => {
-  const newSize = size * scale;
-  if (Platform.OS === 'ios') {
-    return Math.round(PixelRatio.roundToNearestPixel(newSize));
-  } else {
-    return Math.round(PixelRatio.roundToNearestPixel(newSize)) - 2;
-  }
-};
+import normalize from '../../../library/methods/normalize';
 
 export const styles = StyleSheet.create({
   layout: {
