+++ src/library/methods/index.js
@@ -1,10 +1,12 @@
 import {performGoogleAuth, performGoogleAuthSign} from './performAuth';
 import {sortArrayByDate} from './sortArrayByDate';
 import {dateFormatter} from './dateFormatter';
+import normalize from '../methods/normalize';
 
 export {
   performGoogleAuth,
   performGoogleAuthSign,
   sortArrayByDate,
   dateFormatter,
+  normalize,
 };
